Refactor cmd package and buffer channels
