+++ cmd/importer.go
@@ -2,15 +2,9 @@ package main
 
 import (
 	"flag"
-	"fmt"
 	"log"
-	"time"
 
-	"github.com/yixinglu/nebula-importer/pkg/client"
-	"github.com/yixinglu/nebula-importer/pkg/config"
-	"github.com/yixinglu/nebula-importer/pkg/errhandler"
-	"github.com/yixinglu/nebula-importer/pkg/reader"
-	"github.com/yixinglu/nebula-importer/pkg/stats"
+	"github.com/yixinglu/nebula-importer/pkg/cmd"
 )
 
 var configuration = flag.String("config", "", "Specify importer configure file path")
@@ -18,32 +12,7 @@ var configuration = flag.String("config", "", "Specify importer configure file p
 func main() {
 	flag.Parse()
 
-	yaml, err := config.Parse(*configuration)
-	if err != nil {
+	if err := cmd.Run(*configuration); err != nil {
 		log.Fatal(err)
-	}
-
-	now := time.Now()
-	defer func() {
-		fmt.Printf("\nFinish import data, consume time: %.2fs\n", time.Since(now).Seconds())
-	}()
-
-	statsMgr := stats.NewStatsMgr()
-	defer statsMgr.Close()
-
-	clientMgr := client.NewNebulaClientMgr(yaml.NebulaClientSettings, statsMgr.StatsCh)
-	defer clientMgr.Close()
-
-	for _, file := range yaml.Files {
-		clientMgr.InitFile(file)
-
-		errWriter := errhandler.New(file, clientMgr.GetErrChan(), statsMgr.StatsCh)
-		errWriter.InitFile(file, yaml.NebulaClientSettings.Concurrency)
-
-		r := reader.New(file, clientMgr.GetDataChans())
-		r.Read()
-
-		// Wait to finish handle errors
-		<-statsMgr.FileDoneCh
 	}
 }
